Create practical_1_task_2
Task: 
Create a Card object by picking an integer for the value and another one for the suits and try to print it. 

What do you observe? Try to print the value and the suit of the card by accessing the instance lists.

Hint: You may want to go back to the value list and modify it in order to align the index with the displayed value by adding some None elements